@@ -1,10 +1,11 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { BlogArticleComponent } from './blog-article.component';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { of } from 'rxjs';
 import { blogData } from 'src/app/global/data/blog-data';
 import { RouterTestingModule } from '@angular/router/testing';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 declare let process: any;
 let env = process.env.NODE_ENV;
@@ -15,9 +16,10 @@ describe('BlogArticleComponent', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      declarations: [BlogArticleComponent],
-      imports: [HttpClientTestingModule, RouterTestingModule]
-    });
+    declarations: [BlogArticleComponent],
+    imports: [RouterTestingModule],
+    providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+});
     fixture = TestBed.createComponent(BlogArticleComponent);
     component = fixture.componentInstance;
     env = 'production'
